tmp -> main, add blocks(o,z,...)
