+++ src/seoultech/se/tetris/blocks/JBlock.java
@@ -1,14 +1,32 @@
 package seoultech.se.tetris.blocks;
 
-import java.awt.Color;
+import java.awt.*;
 
-public class JBlock extends Block {
-	
-	public JBlock() {
-		shape = new int[][] { 
-				{1, 1, 1},
-				{0, 0, 1}
-		};
-		color = Color.BLUE;
-	}
+public class JBlock extends ParentBlock {
+
+    public JBlock() {
+        color = Color.BLUE;
+        shapes = new Block[][][] {
+                {
+                    {new Block(color), new Block(color), new Block(color)},
+                    {null, null, new Block(color)}
+                },
+                {
+                    {null, new Block(color)},
+                    {null, new Block(color)},
+                    {new Block(color), new Block(color)}
+                },
+                {
+                    { new Block(color), null, null,},
+                    {new Block(color), new Block(color), new Block(color)},
+                },
+                {
+                    {new Block(color), new Block(color)},
+                    {new Block(color), null},
+                    {new Block(color), null},
+                },
+        };
+        shape = shapes[type];
+
+    }
 }
