+++ src/seoultech/se/tetris/blocks/LBlock.java
@@ -1,14 +1,32 @@
 package seoultech.se.tetris.blocks;
 
-import java.awt.Color;
+import java.awt.*;
 
-public class LBlock extends Block {
-	
-	public LBlock() {
-		shape = new int[][] { 
-			{1, 1, 1},
-			{1, 0, 0}
-		};
-		color = Color.ORANGE;
-	}
+public class LBlock extends ParentBlock {
+
+    public LBlock() {
+        color = Color.BLUE;
+
+        shapes = new Block[][][]{
+                {
+                        {new Block(color), new Block(color), new Block(color),},
+                        {new Block(color), null, null},
+                },
+                {
+                        {new Block(color), new Block(color)},
+                        {null, new Block(color),},
+                        {null, new Block(color),},
+                },
+                {
+                        {null, null, new Block(color),},
+                        {new Block(color), new Block(color), new Block(color),},
+                },
+                {
+                        {new Block(color), null,},
+                        {new Block(color), null,},
+                        {new Block(color), new Block(color)},
+                },
+        };
+        shape = shapes[type];
+    }
 }
