+++ src/seoultech/se/tetris/blocks/OBlock.java
@@ -2,13 +2,30 @@ package seoultech.se.tetris.blocks;
 
 import java.awt.*;
 
-public class OBlock extends Block {
+public class OBlock extends ParentBlock {
 
-	public OBlock() {
-		shape = new int[][] { 
-			{1, 1}, 
-			{1, 1}
-		};
-		color = Color.YELLOW;
-	}
+    public OBlock() {
+        color = Color.YELLOW;
+
+        shapes = new Block[][][]{
+                {
+                        {new Block(color), new Block(color)},
+                        {new Block(color), new Block(color)},
+                },
+                {
+                        {new Block(color), new Block(color)},
+                        {new Block(color), new Block(color)},
+                },
+                {
+                        {new Block(color), new Block(color)},
+                        {new Block(color), new Block(color)},
+                },
+                {
+                        {new Block(color), new Block(color)},
+                        {new Block(color), new Block(color)},
+                },
+        };
+
+        shape = shapes[type];
+    }
 }
