+++ src/seoultech/se/tetris/blocks/ZBlock.java
@@ -1,14 +1,67 @@
 package seoultech.se.tetris.blocks;
 
-import java.awt.Color;
+import java.awt.*;
 
-public class ZBlock extends Block {
-	
-	public ZBlock() {
-		shape = new int[][] { 
-			{1, 1, 0},
-			{0, 1, 1}
-		};
-		color = Color.RED;
-	}
+public class ZBlock extends ParentBlock {
+
+    public ZBlock() {
+        color = Color.RED;
+
+        shapes = new Block[][][]{
+                {
+                        {
+                                new Block(color),
+                                new Block(color),
+                                null,
+                        },
+                        {
+                                null,
+                                new Block(color),
+                                new Block(color),
+                        },
+                },
+                {
+                        {
+                                null,
+                                new Block(color),
+                        },
+                        {
+                                new Block(color),
+                                new Block(color),
+                        },
+                        {
+                                new Block(color),
+                                null,
+                        },
+                },
+                {
+                        {
+                                new Block(color),
+                                new Block(color),
+                                null,
+                        },
+                        {
+                                null,
+                                new Block(color),
+                                new Block(color),
+                        },
+                },
+                {
+                        {
+                                null,
+                                new Block(color),
+                        },
+                        {
+                                new Block(color),
+                                new Block(color),
+                        },
+                        {
+                                new Block(color),
+                                null,
+                        },
+                },
+        };
+
+        shape = shapes[type];
+    }
 }
